fix: update glTF types, upstream exporter with uv compat fix
